29/08/2022 15:21

+++ src/main/java/com/example/test/repositories/hotelRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.test.Entities.hotel;
 
-public interface hotelRepository extends JpaRepository<hotel,Integer>{
+public interface hotelRepository extends JpaRepository<hotel,Long>{
     
 }
